fix: enhance brand management with ID validation and retrieval functi…

+++ src/modules/brands/brand.services.ts
@@ -16,11 +16,19 @@ export const createBrand = async (data: brandInput, file: Express.Multer.File) =
 }
 
 export const updateBrand = async (id: number, data: brandInput, file: Express.Multer.File) => {
+    const existingBrand = await brandRepositories.getBrandById(id);
+    if (!existingBrand) {
+        throw new Error("NOT_FOUND_BRAND");
+    }
     const slug = slugify(data.name, { lower: true });
     return await brandRepositories.updateBrand({ id, ...data, slug }, file);
 }
 
 export const deleteBrand = async (id: number) => {
+    const existingBrand = await brandRepositories.getBrandById(id);
+    if (!existingBrand) {
+        throw new Error("NOT_FOUND_BRAND");
+    }
     return await brandRepositories.deleteBrand(id);
 }
